Did refactroing: split PongService
